split options and specification in config, extract spec generation
